test_task/requirements.txt
Django==1.8
django-crispy-forms==1.6.0
XlsxWriter==0.8.5
